requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2==2.8.5
django-rest-auth==0.9.5
django-allauth==0.42.0
django-extensions==3.0.9
django-filter==2.3.0
